move rendering stuff into subsystem
