+++ ne_engine/CMakeLists.txt
@@ -1,13 +1,14 @@
 add_library(
     ne_engine STATIC
 
-        "public/src/ShaderProgram.cpp"
+    "public/src/ShaderProgram.cpp"
     "public/src/Engine.cpp"
     "public/src/Input.cpp"
     "public/src/CharWrapper.cpp"
     "public/src/Triangle.cpp"
     "public/src/JSONLoader.cpp"
+    "public/src/Render.cpp"
 )
 
 target_include_directories(ne_engine PUBLIC public/include PRIVATE private/include)
-target_link_libraries(ne_engine glad glfw glm nlohmann_json::nlohmann_json)
+target_link_libraries(ne_engine glad glfw glm nlohmann_json::nlohmann_json entt)
